code clean-up

+++ inventario-lgpd-backend/src/controllers/users.controller.ts
@@ -1,8 +1,8 @@
-import express, { Request, Response } from "express";
+import { Request, Response } from "express";
+import { validationResult } from "express-validator";
 
 import { BaseUser, User } from "../models/users.model";
 import * as UserService from "../services/users.service";
-import { validationResult } from "express-validator";
 
 export const getUsers = async (req: Request, res: Response) => {
   try {
@@ -20,11 +20,7 @@ export const getUserById = async (req: Request, res: Response) => {
   try {
     const reqUser: User = await UserService.find(id);
 
-    if (reqUser) {
-      return res.status(200).send(reqUser);
-    }
-
-    res.status(404).send("Usuário não encontrado");
+    return res.status(200).send(reqUser);
   } catch (error: any) {
     res.status(500).send(error.message);
   }
@@ -39,14 +35,7 @@ export const registerUser = async (req: Request, res: Response) => {
 
     const receivedUser: BaseUser = req.body;
 
-    const hasUser = await UserService.findByUserName(receivedUser.username);
-
-    if (hasUser) {
-      return res.status(422).send("Já existe um usuário com este nome!");
-    }
-
     const newUser = await UserService.create(receivedUser);
-
     res.status(201).json(newUser);
   } catch (error: any) {
     res.status(500).send(error.message);
@@ -54,21 +43,12 @@ export const registerUser = async (req: Request, res: Response) => {
 };
 
 export const updateUser = async (req: Request, res: Response) => {
-  const id: string = req.params.uid;
-
   try {
+    const id: string = req.params.uid;
     const userUpdate: User = req.body;
 
-    const existingUser: User = await UserService.find(id);
-
-    if (existingUser) {
-      const updatedUser = await UserService.update(id, userUpdate);
-      return res.status(200).json(updatedUser);
-    }
-
-    const newUser = await UserService.create(userUpdate);
-
-    res.status(201).json(newUser);
+    const updatedUser = await UserService.update(id, userUpdate);
+    return res.status(200).json(updatedUser);
   } catch (error: any) {
     res.status(500).send(error.message);
   }
@@ -77,9 +57,15 @@ export const updateUser = async (req: Request, res: Response) => {
 export const removeUser = async (req: Request, res: Response) => {
   try {
     const id: string = req.params.uid;
-    await UserService.remove(id);
 
-    res.status(200).send("Usuário removido com sucesso");
+    const existingUser = await UserService.find(id);
+
+    if (existingUser) {
+      const removedUser = await UserService.remove(id);
+      return res.status(200).json(removedUser);
+    }
+
+    res.status(404).send("Usuário não encontrado");
   } catch (error: any) {
     res.status(500).send(error.message);
   }
@@ -98,5 +84,7 @@ export const loginUser = async (req: Request, res: Response) => {
     }
 
     res.status(200).send("Usuário logado!");
-  } catch (error) {}
+  } catch (error: any) {
+    res.status(500).send(error.message);
+  }
 };
